refactor: 로그인 auth 도메인 리팩토링
