--- src/components/Login/index.ts
@@ -1,4 +1,0 @@
-export { default as HisnetLoginButton } from './HisnetLoginButton';
-export { default as LoginSection } from './LoginSection';
-export { default as LogoSection } from './LogoSection';
-export { default as NoticeCard } from './NoticeCard';
