+++ src/components/_common/Drawer/LogoutSection.tsx
@@ -2,10 +2,11 @@ import { LogoutIcon } from '@/assets';
 import { Text } from '@/components';
 import Flex from '@/components/_common/Flex/Flex';
 import { ROUTE_PATH } from '@/constants/routePath';
-import { usePostLogoutMutation } from '@/hooks/queries';
 import { getOpacityColor } from '@/utils/getOpacityColor';
 import { styled, useTheme } from '@mui/material';
 import { useNavigate } from 'react-router-dom';
+
+import { usePostLogoutMutation } from '@/pages/LoginPage/hooks/usePostLogoutMutation';
 
 const LogoutSection = () => {
   const theme = useTheme();
