+++ src/hooks/index.ts
@@ -2,5 +2,4 @@ export { default as useFile } from './useFile';
 export { default as useFileWithType } from './useFileWithType';
 export { default as useInput } from './useInput';
 export { default as useInputWithValidate } from './useInputWithValidate';
-export { default as useLogin } from './useLogin';
 export { default as useOpenModal } from './useOpenModal';
