--- src/hooks/queries/index.ts
@@ -1,4 +1,0 @@
-export { default as usePostLogoutMutation } from './usePostLogoutMutation';
-
-export * from './useGetUserInfoQuery';
-export * from './usePostLoginMutation';
