+++ src/pages/LoginPage/apis/auth.ts
@@ -1,7 +1,8 @@
 import { ENDPOINT } from '@/apis/endPoint';
 import { http } from '@/apis/http';
-import { AUTH_ACCESS_KEY } from '@/constants/auth';
-import { AuthRequest, UserResponse } from '@/types/auth';
+
+import { AUTH_ACCESS_KEY } from '../constants/auth';
+import { AuthRequest, UserResponse } from '../types/auth';
 
 export const postLogin = async ({ token }: Omit<AuthRequest, 'accessKey'>) => {
   const response = await http.post<AuthRequest, UserResponse>(
