+++ src/pages/LoginPage/hooks/usePostLoginMutation.ts
@@ -1,9 +1,10 @@
-import { postLogin } from '@/apis/auth';
 import { TOAST_MESSAGES } from '@/constants/toastMessage';
 import { amplitudeInitializer } from '@/service/amplitude/amplitudeInitializer';
 import { useAuthStore } from '@/stores';
 import { useMutation } from '@tanstack/react-query';
 import { toast } from 'react-toastify';
+
+import { postLogin } from '../apis/auth';
 
 export const usePostLoginMutation = () => {
   const { init } = amplitudeInitializer();
