+++ src/pages/LoginPage/types/auth.ts
