+++ src/pages/MyPage/components/GithubGraphSection.tsx
@@ -1,7 +1,8 @@
 import { Flex } from '@/components';
-import { useGetUserInfoQuery } from '@/hooks/queries';
 import { boxShadow } from '@/styles/common';
 import { styled, useTheme } from '@mui/material';
+
+import { useGetUserInfoQuery } from '@/pages/LoginPage/hooks/useGetUserInfoQuery';
 
 export const GithubGraphSection = () => {
   const theme = useTheme();
