+++ src/pages/ScholarshipApplyPage/constants/personalInfoConsent.ts
@@ -1,4 +1,5 @@
 import personalInfoConsentData from '@/data/personalInfoConsentData.json';
-import { Consent } from '@/types/consent';
+
+import { Consent } from '../types/consent';
 
 export const personalInfoConsent: Consent = personalInfoConsentData;
