+++ src/pages/ScholarshipApplyPage/types/consent.ts
